+++ src/helpers/custom.ts
@@ -1,9 +1,9 @@
-import { CUSTOM } from '../constants';
-import type { Custom } from '../types';
+import { CUSTOM } from "../constants";
+import type { Custom } from "../types";
 
 const custom = <T = any>(value?: T) =>
   ({
     [CUSTOM]: value,
   }) as Custom<T>;
 
-export default custom;
+export { custom };