+++ src/helpers/sv.ts
@@ -1,4 +1,4 @@
-import type { Custom, StateValue } from '../types';
+import type { Custom, StateValue } from "../types";
 
 const sv = {} as Custom<StateValue>;
-export default sv;
+export { sv };