+++ src/helpers/tuple.ts
@@ -3,4 +3,4 @@ import type { NotReadonly, ObjectT } from "../types";
 const tuple = <const T extends [ObjectT, ...ObjectT[]]>(...values: T) =>
   values as NotReadonly<T>;
 
-export default tuple;
+export { tuple };