+++ src/index.ts
@@ -1,4 +1,4 @@
-export * from './constants';
-export * from './helpers';
-export * from './type';
-export * from './types';
+export * from "./constants";
+export * as helpers from "./helpers";
+export * from "./type";
+export * from "./types";